Refactor new map implementation into reusable component
