+++ webclient/app/src/app/commons/geographicalMaps/BaseMap.jsx
@@ -1,0 +1,37 @@
+import {MapboxOverlay} from '@deck.gl/mapbox';
+import 'maplibre-gl/dist/maplibre-gl.css';
+import {Map, useControl, AttributionControl, NavigationControl} from 'react-map-gl/maplibre';
+
+
+function DeckGLOverlay(props) {
+  const overlay = useControl(() => new MapboxOverlay(props));
+  overlay.setProps(props);
+  return null;
+}
+
+function BaseMap(props) {
+    const {
+        viewState,
+        onViewStateChange,
+        layers = [],
+        getTooltip = null,
+    } = props;
+
+    return (
+        <div style={{width: "100vw", height: "100vh", position: "fixed", top: 0, left: 0 }}>
+            <Map
+                initialViewState={viewState}
+                mapStyle="https://tiles.openfreemap.org/styles/positron"
+                onMove={evt => onViewStateChange(evt.viewState)}
+                {...viewState}
+            >
+                <DeckGLOverlay
+                    layers={layers}
+                    getTooltip={getTooltip}
+                />
+            </Map>
+        </div>
+    );
+}
+
+export default BaseMap;
